+++ src/main/java/br/edu/fatecsjc/lgnspringapi/service/MarathonsService.java
@@ -2,7 +2,6 @@ package br.edu.fatecsjc.lgnspringapi.service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,58 +16,65 @@ import jakarta.transaction.Transactional;
 
 @Service
 public class MarathonsService {
+
     @Autowired
-    private MarathonsRepository marathonsRepository;   
+    private MarathonsRepository marathonsRepository;
+
     @Autowired
     private MemberRepository memberRepository;
+
     @Autowired
     private MarathonsConverter marathonsConverter;
-    
+
     public void addMarathon(Member member, Marathons marathon) {
         if (member.getMarathons() == null) {
             member.setMarathons(new ArrayList<>());
         }
         member.getMarathons().add(marathon);
     }
-    
-    public List<MarathonsDTO> getAll(){
-    	return marathonsConverter.convertToDto(marathonsRepository.findAll());
+
+    public List<MarathonsDTO> getAll() {
+        return marathonsConverter.convertToDto(marathonsRepository.findAll());
     }
-    
+
     public MarathonsDTO findById(Long id) {
-        return marathonsConverter.convertToDto(marathonsRepository.findById(id).get());
+        return marathonsConverter.convertToDto(marathonsRepository.findById(id).orElse(null));
     }
-    
+
     @Transactional
     public MarathonsDTO save(Long id, MarathonsDTO dto) {
         Marathons entity = marathonsRepository.findById(id)
-            .orElseThrow(() -> new RuntimeException("Maratona não encontrada!"));
+            .orElse(null);
 
-        memberRepository.deleteMembersByMarathons(entity);
-        entity.getMembers().clear();
+        if (entity != null) {
+            memberRepository.deleteMembersByMarathons(entity);
+            entity.getMembers().clear();
 
-        Marathons marathonsToSaved = marathonsConverter.convertToEntity(dto, entity);
+            Marathons marathonsToSaved = marathonsConverter.convertToEntity(dto, entity);
 
-        if (marathonsToSaved.getMembers() != null) {
-            marathonsToSaved.getMembers().forEach(member -> {
-                if (member.getMarathons() == null) {
-                    member.setMarathons(new ArrayList<>()); 
-                }
-                member.getMarathons().add(marathonsToSaved); 
-            });
+            if (marathonsToSaved.getMembers() != null) {
+                marathonsToSaved.getMembers().forEach(member -> {
+                    if (member.getMarathons() == null) {
+                        member.setMarathons(new ArrayList<>());
+                    }
+                    member.getMarathons().add(marathonsToSaved);
+                });
+            }
+
+            Marathons marathonReturned = marathonsRepository.save(marathonsToSaved);
+            return marathonsConverter.convertToDto(marathonReturned);
         }
 
+        return null;
+    }
+
+    public MarathonsDTO save(MarathonsDTO dto) {
+        Marathons marathonsToSaved = marathonsConverter.convertToEntity(dto);
         Marathons marathonReturned = marathonsRepository.save(marathonsToSaved);
         return marathonsConverter.convertToDto(marathonReturned);
     }
-    
-    public MarathonsDTO save(MarathonsDTO dto) {
-    	Marathons marathonsToSaved = marathonsConverter.convertToEntity(dto);
-    	Marathons marathonReturned = marathonsRepository.save(marathonsToSaved);
-    	return marathonsConverter.convertToDto(marathonReturned);
-    }
-    
+
     public void delete(Long id) {
-    	marathonsRepository.deleteById(id);
+        marathonsRepository.deleteById(id);
     }
 }
